Added development environment variable to default gulp task. Added Hapi plugins: hapi-router, blipp, disinfect, hapi-alive, hapi-dev-errors. Added controllers directory and refactored controller logic out of the routes into separate controllers. Implemented cache-busting for the main application.css file based on the version number in package.json
