+++ index.js
@@ -1,5 +1,12 @@
 const Hapi = require('hapi')
+const HapiRouter = require('hapi-router')
+const Blipp = require('blipp')
+const Disinfect = require('disinfect')
+const HapiAlive = require('hapi-alive')
+const HapiDevErrors = require('hapi-dev-errors')
 const server = new Hapi.Server()
+
+const environment = process.env.NODE_ENV
 
 server.connection({
   port: 8000
@@ -32,34 +39,56 @@ server.register([
   require('inert'),
   require('vision'),
   {
-    register: require('disinfect'),
+    // Plugin to automatically load the routes
+    register: HapiRouter,
+    options: {
+      routes: './src/routes/*.js'
+    }
+  }, {
+    // Plugin to display the routes table to console at startup
+    register: Blipp,
+    options: {}
+  }, {
+    // Plugin to prevent CSS attack by applying Google's Caja HTML Sanitizer on route query, payload, and params
+    register: Disinfect,
     options: {
       disinfectQuery: true,
       disinfectParams: true,
       disinfectPayload: true
     }
-  }],
-  (err) => {
-    if (err) {
-      throw err
+  }, {
+    // Plugin providing a health route for the server
+    register: HapiAlive,
+    options: {
+      path: '/health',
+      healthCheck: function (server, callback) {
+            // TODO: Here you should preform your health checks
+            // If something went wrong provide the callback with an error
+        callback()
+      }
     }
+  }, {
+    // Plugin to return an error view for web request. Only used in development
+    register: HapiDevErrors,
+    options: {
+      showErrors: environment === 'development'
+    }
+  }], (err) => {
+  if (err) {
+    throw err
+  }
 
-    // Load views
-    server.views(require('./src/views'))
-
-    // TODO load in routes using wildcard
-    // Load routes
-    server.route(require('./src/routes/public'))
-    server.route(require('./src/routes/default'))
-    server.route(require('./src/routes/error'))
-    server.route(require('./src/routes/taskList'))
-    server.route(require('./src/routes/site'))
-  })
+  // Load views
+  server.views(require('./src/views'))
+})
 
 // Start the server
 server.start((err) => {
   if (err) {
     throw err
+  }
+  if (environment) {
+    console.info('Server running in environment: [' + environment + ']')
   }
   console.info('Server running at:', server.info)
 })
