+++ src/routes/default.js
@@ -1,21 +1,5 @@
-// TODO confirm which version of UUID to use
-const uuid4 = require('uuid/v4')
-
 module.exports = [{
   method: 'GET',
   path: '/',
-  handler: (request, reply) => {
-    const context = {
-      pageTitle: 'Waste Permits'
-    }
-
-    // Generate a token
-    const token = uuid4()
-    console.log('Generated token: ' + token)
-
-    reply
-      .view('index', context)
-      .unstate('session')                   // Delete existing session cookie
-      .state('session', { token: token })   // Create new session cookie
-  }
+  handler: require('../controllers/default')
 }]
