+++ src/views/partials/common/head.html
@@ -1,1 +1,1 @@
-<link href="public/stylesheets/application.css" media="screen" rel="stylesheet" />
+<link href="public/stylesheets/application.css?##APP_VERSION##" media="screen" rel="stylesheet" />
